rename api folder to queries
