+++ next-app/pages/admin/honorarios.tsx
@@ -5,7 +5,7 @@ import { useSession } from 'next-auth/react'
 import SendButton from '@/components/SendButton'
 import Link from 'next/link'
 import { FaChevronLeft } from 'react-icons/fa'
-import { useHonorarios } from '@/api/admin/honorarios'
+import { useHonorarios } from '@/queries/admin/honorarios'
 import { editHonorarios } from '@/schemas/honorarios'
 import useFormErrors from '@/hooks/useFormErrors'
 import { BiLoaderAlt } from 'react-icons/bi'
