+++ next-app/queries/admin/cursos.ts
